+++ packages/capacitor/src/generators/capacitor-project/schema.d.ts
@@ -1,0 +1,12 @@
+export interface CapacitorGeneratorSchema {
+  project: string;
+  appId: string;
+  appName?: string;
+  webDir?: string;
+  npmClient: string;
+}
+
+export interface NormalizedSchema extends CapacitorGeneratorSchema {
+  projectRoot: string;
+  pathToRoot: string;
+}