+++ packages/capacitor/src/schematics/add-plugin/schema.d.ts
@@ -1,9 +1,0 @@
-export interface AddPluginSchematicSchema {
-  plugin: string;
-  version: string;
-  project: string;
-}
-
-export interface NormalizedSchema extends AddPluginSchematicSchema {
-  projectRoot: string;
-}